% Generated by roxygen2 (4.0.1): do not edit by hand
\name{add_cat}
\alias{add_cat}
\title{Add a kitty to your ggplot2 plot}
\usage{
add_cat(..., lighten = 0.7, bw = TRUE)
}
\arguments{
\item{...}{Arguments passed to \code{\link{get_cat}}, i.e. \code{height} and \code{width}}

\item{lighten}{Number between 0 (no adjustment) and 1 (all white) to lighten the image}

\item{bw}{Logical, the image should be displayed in grey-scale (\code{FALSE} = full colour)}
}
\description{
Grabs a random cat image from \url{http://theoldreader.com/kittens} and creates a
\code{\link{annotation_raster}} layer to add to your ggplot2 plot.
}
\examples{
library(ggplot2)
ggplot(mpg, aes(cty, hwy)) +
 add_cat() +
 geom_point()

# In glorious kitty colours
ggplot(mpg, aes(cty, hwy)) +
 add_cat(bw = FALSE) +
 geom_point()
}
\keyword{cats}

