% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{add_singletons_ts}
\alias{add_singletons_ts}
\title{Add singletons to a dataset}
\usage{
add_singletons_ts(spab_ts, use_max = TRUE)
}
\arguments{
\item{spab_ts}{the dataset over the timeseries, col timestep}

\item{use_max}{use max?}
}
\value{
dataset plus singletons
}
\description{
Add singletons to a dataset
}
