% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_kmers.R
\name{all_kmers}
\alias{all_kmers}
\title{Generate all kmers of length k}
\usage{
all_kmers(k)
}
\arguments{
\item{k}{the length of kmers to generate}
}
\value{
character vector of kmers lexicographically
}
\description{
Generate all kmers of length k
}
\examples{
all_kmers(3)
}
