% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_function.R
\name{cat_function}
\alias{cat_function}
\title{A Cat function}
\usage{
cat_function(love = TRUE)
}
\arguments{
\item{love:}{do you love cats? dafult is TRUE}
}
\description{
This function allows to show ones love for cats.
At times it may be difficult to express the same by anyother means
}
\examples{
cat_function()
}
\keyword{cats}
