% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catlove.R
\name{cat_love}
\alias{cat_love}
\title{Do you love cats?}
\usage{
cat_love(love = TRUE)
}
\arguments{
\item{Insert}{TRUE or FALSE to answer if you love cats}
}
\value{
Shows a quote after answering whether or not you love cats.
}
\description{
Answer TRUE or FALSE when answering if you love cats.
}
