% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cats-package.R
\docType{package}
\name{cats-package}
\alias{cats-package}
\title{Power calculation for a joint analysis of a two-stage case control design
for SNP data}
\description{
Power calculation for a joint analysis of a two-stage case control design
for SNP data.
}
\examples{
# calculate the power under a multiplicative model using a
# two stage design and assuming a relative risk of 1.5
cats(
  freq = 0.2,
  ncases = 500, ncases2 = 500,
  ncontrols = 1000, ncontrols2 = 1000,
  risk = 1.5, multiplicative = 1
)
}
\references{
Skol AD, Scott LJ, Abecasis GR, Boehnke M: Joint analysis is
more efficient than replication-based analysis for two-stage genome-wide
association studies. Nat Genet 38: 209-213, 2006.
}
\seealso{
\url{http://www.sph.umich.edu/csg/abecasis/CaTS/}
}
\author{
Anders Albrechtsen <albrecht@binf.ku.dk>
}
