% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATS.R
\name{ContextAgnosticExpansion}
\alias{ContextAgnosticExpansion}
\title{STEP 1: Context agnostic network expansion}
\usage{
ContextAgnosticExpansion(expressionList, PPI, pCut = 0.05, FCCut = 3)
}
\arguments{
\item{expressionList}{List of gene expression for signatures and the rest
of the genes obtained from the previous step, STEP 0
with getSignatureGenes' function.}

\item{PPI}{Data frame formatted protein-protein interaction (PPI) networks}

\item{pCut}{Cut-off for the fraction of genes in the signature list
connected with the gene g, which is not in the signature list.
By default, it is 0.05}

\item{FCCut}{Fold change (FC) cut-off for this step. By default, it is 3}
}
\value{
Returns the expanded signature and the rest of the genes' expression
        matrix as a list
}
\description{
STEP 1: Context agnostic network expansion
}
\examples{
step0 <- getSignatureGenes(signature, gene_expression)
step1 <- ContextAgnosticExpansion(expressionList = step0, PPI = human_PPIN,
                                  pCut = 0.001, FCCut = 0.2)

}
