% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATS.R
\name{ContextSpecificExpansion}
\alias{ContextSpecificExpansion}
\title{STEP 2: Context-specific network expansion}
\usage{
ContextSpecificExpansion(
  expressionList,
  corCutOff = 0.3,
  FCCut = 3,
  fracCut = 0.05,
  nThread = 5
)
}
\arguments{
\item{expressionList}{List of gene expression for signatures and the rest
of the genes obtained from the previous step, STEP 1
with 'ContextAgnosticExpansion' function.}

\item{corCutOff}{Alpha cut-off. It defines an edge if the correlation is
bigger than the alpha. Here we term alpha as a ratio
between the fraction of genes in Sig-PPI-Expanded
connected with g and the fraction of genes genome-wide
connected with g.}

\item{FCCut}{fold change (FC) cut-off for this step. By default, it is 3}

\item{fracCut}{The fraction cut-off for the genes in the previously expanded
gene set that are connected with gene g which is not in the
expanded gene list.}

\item{nThread}{The number thread that will be used in the analysis}
}
\value{
Returns expanded gene signatures in the given context list
}
\description{
STEP 2: Context-specific network expansion
}
\examples{
step0 <- getSignatureGenes(signature, gene_expression)
step1 <- ContextAgnosticExpansion(step0, human_PPIN, pCut = 0.01)
step2 <- ContextSpecificExpansion(step1)


}
