% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve.cats.R
\name{curve.cats}
\alias{curve.cats}
\title{Curves plot for cats object}
\usage{
curve.cats(RR, MAF, rep = 1536, SNPs = 1e+06, ncases, ncontrols,
  ncases2, ncontrols2, alpha = 0.05/SNPs, ...)
}
\arguments{
\item{RR}{description}

\item{MAF}{description}

\item{rep}{description}

\item{SNPs}{description}

\item{ncases}{description}

\item{ncontrols}{description}

\item{ncases2}{description}

\item{ncontrols2}{description}

\item{alpha}{description}

\item{...}{description}
}
\description{
Curves plot for cats object
}
