% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DE_genes}
\alias{DE_genes}
\title{List of genes that are differentially expressed in TCGA brain RNASeq}
\format{
Not Available
}
\source{
\url{https://portal.gdc.cancer.gov/}
}
\usage{
DE_genes
}
\description{
List of genes that are differentially expressed in TCGA brain RNASeq
}
\examples{
data(DE_genes, package = 'CATS')
}
\keyword{datasets}
