% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dog_function.R
\name{dog_function}
\alias{dog_function}
\title{Expresses a well known love of dogs.}
\usage{
dog_function(love = T)
}
\arguments{
\item{love}{defaults to TRUE for obvious reasons}
}
\description{
Expresses a well known love of dogs.
}
\examples{
dog_function(F)
}
