% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_fs_ts.R
\name{fs_ts_wrapper}
\alias{fs_ts_wrapper}
\title{TS Sampling wrapper}
\usage{
fs_ts_wrapper(spab, nsamples = 10, seed = 1, p_table = NULL)
}
\arguments{
\item{spab}{df}

\item{nsamples}{nsamples}

\item{seed}{set a seed}

\item{p_table}{path to p table}
}
\value{
df of dis of obs vectors
}
\description{
TS Sampling wrapper
}
