% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_expression}
\alias{gene_expression}
\title{Subset of gene expressions in TCGA brain RNASeq data}
\format{
Not Available
}
\source{
\url{https://portal.gdc.cancer.gov/}
}
\usage{
gene_expression
}
\description{
Subset of gene expressions in TCGA brain RNASeq data
}
\examples{
data(gene_expression, package = 'CATS')
}
\keyword{datasets}
