% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_directional_pairs.R
\name{generate_pairings}
\alias{generate_pairings}
\title{Generate random directional pairs from a vector of names (nodes)}
\usage{
generate_pairings(nodes)
}
\arguments{
\item{nodes}{}
}
\value{
dataframe with pairs as rows
}
\description{
Generate random directional pairs from a vector of names (nodes)
}
\examples{
generate_pairings(c('Sporty', 'Ginger','Scary','Baby','Posh'))
}
