% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_function.R
\name{get_love_state}
\alias{get_love_state}
\title{Access Function}
\usage{
get_love_state()
}
\description{
This function allows you to check on the state of love for cats.
}
\examples{
get_love_state()
}
\keyword{cats}
