% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcdb.R
\name{get_mcdb_spab}
\alias{get_mcdb_spab}
\title{Load spab table for a MCDB site}
\usage{
get_mcdb_spab(mcdb_site)
}
\arguments{
\item{mcdb_site}{which site}
}
\value{
spab table
}
\description{
Load spab table for a MCDB site
}
