% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{get_statevars_ts}
\alias{get_statevars_ts}
\title{Get statevars on a TS}
\usage{
get_statevars_ts(spab)
}
\arguments{
\item{spab}{from make_spab}
}
\value{
dataframe with cols site singletons dat source sim timestep s0 n0
}
\description{
Get statevars on a TS
}
