% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{get_toy_bbs_data}
\alias{get_toy_bbs_data}
\title{Get toy BBS data}
\usage{
get_toy_bbs_data(route = 1, region = 11, from_retriever = F)
}
\arguments{
\item{route}{int route}

\item{region}{int region}

\item{from_retriever}{t/f}
}
\value{
MATSS packaged dataset
}
\description{
Load one of 3 BBS route-regions packaged with \code{cats}. For tests and method development, avoids needing MATSS fully functional.
}
\details{
Currently, the options are (rt/rg): 1/11, 268/8, 314/27. These are datasets 1, 410, and 1977 if you run MATSS::build_bbs_datasets_plan. Those are arbitrary numbers chosen to be memorable and disperseed throughout the database (1, our parking space, the first year of Portal).
}
