% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATS.R
\name{getCorr}
\alias{getCorr}
\title{PREPARATION: Finds the correlation cut-off for the STEP 2}
\usage{
getCorr(
  geneExp,
  sizeOfSampling = 2000,
  quantileCut = 0.99,
  samplingGenes = TRUE
)
}
\arguments{
\item{geneExp}{Gene by sample expression matrix. The first column of this
matrix should be gene names.}

\item{sizeOfSampling}{Sampling Size. By default, it is 2000}

\item{quantileCut}{The cut-off that is based on the percentile of randomly
sampled gene pairs. By default, it is 0.99.}

\item{samplingGenes}{Binary variable whether sampling should be based on
genes or samples.This variable should be set to the
TRUE for sampling based on genes. Otherwise, it
should be FALSE}
}
\value{
The correlation cut-off threshold based on the desired quantile
}
\description{
PREPARATION: Finds the correlation cut-off for the STEP 2
}
\examples{
corr_cut <- getCorr(geneExp = gene_expression, sizeOfSampling = 100,
                    quantileCut = 0.95)

}
