% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATS.R
\name{getDEgenes}
\alias{getDEgenes}
\title{STEP 3: Pruning}
\usage{
getDEgenes(geneList, DESet)
}
\arguments{
\item{geneList}{List of gene expression for signatures and the rest of the
genes obtained from the previous step, STEP 2 with
'ContextSpecificExpansion' function.}

\item{DESet}{Differentially expressed gene set or list of interest in
character format}
}
\value{
Returns pruned gene list
}
\description{
STEP 3: Pruning
}
\examples{
step0 <- getSignatureGenes(signature, gene_expression)
step1 <- ContextAgnosticExpansion(step0, human_PPIN, pCut = 0.01)
step2 <- ContextSpecificExpansion(step1)
step3 <- getDEgenes(geneList = step2, DESet = DE_genes)

}
