% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATS.R
\name{getSignatureGenes}
\alias{getSignatureGenes}
\title{STEP 0: Separates the signature gene expression and the rest of the gene
expression}
\usage{
getSignatureGenes(signature, geneExp)
}
\arguments{
\item{signature}{Data frame formatted signature gene list}

\item{geneExp}{Gene by sample expression matrix or SummarizedExperiment
that includes both gene names and gene expression profiles.
The first column of the expression matrix should be gene
names. For SummarizedExperiment, gene names must be provided
 as rownames(geneExp)}
}
\value{
Returns signature and rest of the genes' expression matrix as a
       list
}
\description{
STEP 0: Separates the signature gene expression and the rest of the gene
expression
}
\examples{
step0 <- getSignatureGenes(signature, gene_expression)

}
