% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmissing_data.R
\name{gg_missing}
\alias{gg_missing}
\title{Missing data plot}
\usage{
gg_missing(df = df)
}
\arguments{
\item{df}{dataframe to check for missing data}
}
\value{
ggplot object
}
\description{
Missing data plot
}
