% Generated by roxygen2 (4.0.1): do not edit by hand
\name{here_kitty}
\alias{here_kitty}
\title{Call a kitty to your R session}
\usage{
here_kitty(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{get_cat}}, i.e. \code{height} and \code{width}}
}
\description{
Grabs a random cat image from \url{http://theoldreader.com/kittens} and displays it.
}
\examples{
here_kitty()
}
\keyword{cats}

