% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_PPIN}
\alias{human_PPIN}
\title{Human PPI network}
\format{
Not Available
}
\source{
\url{http://202.97.205.69/CancerSEA/}
}
\usage{
human_PPIN
}
\description{
Human PPI network
}
\examples{
data(signature)
}
\keyword{datasets}
