% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.cats.R
\name{lines.cats}
\alias{lines.cats}
\title{Line plots for cats object}
\usage{
\method{lines}{cats}(x, type = "Replication", col = NULL, lty = 2, ...)
}
\arguments{
\item{x}{description}

\item{type}{description}

\item{col}{description}

\item{lty}{description}

\item{...}{description}
}
\description{
Line plots for cats object
}
