% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{make_spab}
\alias{make_spab}
\title{Make spab table from MATSS dataset}
\usage{
make_spab(matssdat, datname)
}
\arguments{
\item{matssdat}{dataset as imported from MATSS}

\item{datname}{name from drake}
}
\value{
spab table with cols dat site singletons source sim timestep rank species abund
}
\description{
Make spab table from MATSS dataset
}
