% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.supercats.R
\name{plot.supercats}
\alias{plot.supercats}
\title{plot supercats}
\usage{
\method{plot}{supercats}(x, type = "Joint", file = "power.pdf",
  col = NULL, main = paste("POWER N=", x$ncases, ":", x$ncontrols, ",",
  x$ncases2, ":", x$ncontrols2, " rep=", x$rep, sep = ""), ...)
}
\arguments{
\item{x}{description}

\item{type}{description}

\item{file}{description}

\item{col}{description}

\item{main}{description}

\item{...}{description}
}
\description{
plot supercats
}
