% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcdb.R
\name{prep_mcdb_data}
\alias{prep_mcdb_data}
\title{Prepare MCDB data}
\usage{
prep_mcdb_data(from_MATSS = FALSE, save = FALSE)
}
\arguments{
\item{from_MATSS}{use import_retriever_data?}

\item{save}{save to inst?}
}
\value{
if save = F, returns MCDB communities prepped
}
\description{
Prepare MCDB data
}
