% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cats.R
\name{print.cats}
\alias{print.cats}
\title{Print the calculated power}
\usage{
\method{print}{cats}(x, ...)
}
\arguments{
\item{x}{a cats object}

\item{...}{Dynamic parameter for the values of additional parameters for the
summary method}
}
\description{
Print the calculated power
}
\seealso{
cats
}
