% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATS.R
\name{runCATS}
\alias{runCATS}
\title{The pipeline for the CATS.}
\usage{
runCATS(
  signature,
  geneExp,
  PPI,
  DESet,
  pCut = 0.05,
  step1_FCCut = 3,
  corCutOff = 0.3,
  step2_FCCut = 3,
  fracCut = 0.05,
  nThread = 5
)
}
\arguments{
\item{signature}{Data frame formatted signature gene list}

\item{geneExp}{Gene by sample expression matrix. The first column of this
matrix should be gene names.}

\item{PPI}{Data frame formatted protein-protein interaction (PPI) networks}

\item{DESet}{Differentially expressed gene set or list of interest in
character format}

\item{pCut}{Cut-off for the fraction of genes in the signature list
connected with the gene g, which is not in the signature list.
By default, it is 0.05. This parameter is for
ContextAgnosticExpansion step.}

\item{step1_FCCut}{Fold change (FC) cut-off for the Step 1. By default,
it is 3}

\item{corCutOff}{Alpha cut-off. It defines an edge if the correlation is
bigger than the alpha. Here we term alpha as a ratio
between the fraction of genes in Sig-PPI-Expanded
connected with g and the fraction of genes genome-wide
connected with g.}

\item{step2_FCCut}{fold change (FC) cut-off for the step 2.
By default, it is 3}

\item{fracCut}{The fraction cut-off for the genes in the previously expanded
gene set that are connected with gene g which is not in the
expanded gene list.}

\item{nThread}{The number thread that will be used in the analysis}
}
\value{
Gene set
}
\description{
The pipeline for the CATS.
}
\examples{
step3 <- runCATS(signature = signature,
                geneExp = gene_expression,
                PPI = human_PPIN,
                DESet = DE_genes,
                pCut = 0.01)

}
