% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{signature}
\alias{signature}
\title{List of signature genes}
\format{
Not Available
}
\source{
\url{http://202.97.205.69/CancerSEA/}
}
\usage{
signature
}
\description{
List of signature genes
}
\examples{
data(signature)
}
\keyword{datasets}
