% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trial.R
\name{simulate_trial}
\alias{simulate_trial}
\title{Simulates the cohort trial.}
\usage{
simulate_trial(
  n_int,
  n_fin,
  cohorts_start = 1,
  rr_comb,
  rr_mono = 1,
  rr_back = 1,
  rr_plac,
  rr_transform,
  random_type = NULL,
  trial_struc = "all_plac",
  random = FALSE,
  prob_comb_rr = NULL,
  prob_mono_rr = c(1),
  prob_back_rr = c(1),
  prob_plac_rr = NULL,
  prob_rr_transform = prob_rr_transform,
  stage_data = TRUE,
  cohort_random = NULL,
  cohorts_max = 4,
  sr_drugs_pos = 1,
  sr_pats = cohorts_max * (n_fin + 3 * cohorts_max),
  sr_first_pos = FALSE,
  cohort_offset = 0,
  sharing_type = "all",
  safety_prob = 0,
  cohorts_sim = Inf,
  missing_prob = 0,
  cohort_fixed = NULL,
  accrual_type = "fixed",
  accrual_param = 9,
  bio_lag = 12,
  hist_lag = 48,
  time_trend = time_trend,
  ...
)
}
\arguments{
\item{n_int}{Sample size per cohort to conduct interim analysis}

\item{n_fin}{Sample size per cohort at final}

\item{cohorts_start}{Number of cohorts to start the platform with}

\item{rr_comb}{Response rates of combination therapies // high dose}

\item{rr_mono}{Response rate of mono therapies // medium dose}

\item{rr_back}{Response rates of backbone arms  // low dose}

\item{rr_plac}{Response rate of the placebo // SoC}

\item{rr_transform}{Function transforming all the above response rates to a vector of four
probabilities for the multinomial simulation
First element is probability of both failures.
Second element is probability of biomarker success and histology failure.
Third element is probability of biomarker failure and histology success.
Fourth element is probability of both success.}

\item{random_type}{How should the response rates be drawn randomly? Options are:

                            "absolute": Specify absolute response rates that will be drawn with a certain probability

                            "risk_difference": Specify absolute response rates for placebo which will be drawn randomly,
                            plus specify vectors for absolute treatment effects of mono therapies over placebo and
                            for combo over the mono therapies.

                            "risk_ratio": Specify absolute response rates for placebo which will be drawn randomly, plus specify vectors
                            for relative treatment effects of mono therapies over placebo and for combo over the mono therapies.

                            "odds_ratios": Specify response rate for placebo, specify odds-ratios for mono therapies
                            (via rr_back and rr_mono) and respective probabilities.
                            On top, specify interaction for the combination therapy via rr_comb with prob_rr_comb.

                            Set: odds_combo = odds_plac * or_mono1 * or_mono2 * rr_comb.
                            If rr_comb > 1 -> synergistic, if rr_comb = 1 -> additive. If rr_comb < 1 -> antagonistic.
                            Default is "NULL".}

\item{random}{Should the response rates of the arms be randomly drawn from rr_exp? Default is FALSE.}

\item{prob_comb_rr}{If random == TRUE, what are the probabilities with which the elements of rr_comb should be drawn?}

\item{prob_mono_rr}{If random == TRUE, what are the probabilities with which the elements of rr_mono should be drawn?}

\item{prob_back_rr}{If random == TRUE, what are the probabilities with which the elements of rr_back should be drawn?}

\item{prob_plac_rr}{If random == TRUE, what are the probabilities with which the elements of rr_plac should be drawn?}

\item{prob_rr_transform}{If random == TRUE, what are the probabilities with which the elements of rr_transform should be drawn?}

\item{stage_data}{Should individual stage data be passed along? Default is TRUE}

\item{cohort_random}{If not NULL, indicates that new arms/cohorts should be randomly started.
For every timestep, there is a cohort_random probability that a new cohort will be started.}

\item{cohorts_max}{Maximum number of cohorts that are allowed to be added throughout the trial}

\item{sr_drugs_pos}{Stopping rule for successful experimental arms; Default = 1}

\item{sr_pats}{Stopping rule for total number of patients; Default = cohorts_max * n_fin + error term based on randomization}

\item{sr_first_pos}{Stopping rule for first successful cohort;
if TRUE, after first cohort was found to be successful, no further cohorts will be included
but cohorts will finish evaluating, unless other stopping rules reached prior.
Default is FALSE.}

\item{cohort_offset}{Minimum number of time between adding new cohorts}

\item{sharing_type}{Which backbone and placebo data should be used for arm comparisons; Default is "all".
Another option is "concurrent" or "dynamic" or "cohort".}

\item{safety_prob}{Probability for a random stopping after every patient}

\item{cohorts_sim}{Maximum number of cohorts that can run simultaneously}

\item{missing_prob}{Probability for a missing value at final (independent of treatment)}

\item{cohort_fixed}{If not NULL, fixed timesteps after which a cohort will be included}

\item{accrual_type}{Type of patient accrual; choices are "fixed", "poisson" or "exponential"}

\item{accrual_param}{Parameter used for patient accrual}

\item{bio_lag}{Time until biomarker outcome is observed}

\item{hist_lag}{Time until histology outcome is observed}

\item{time_trend}{Additive term by which response rates increase at every time step}

\item{...}{Further arguments to be passed to decision function, such as decision making criteria}

\item{design_type}{Either "combination" or "doses"}

\item{arms_per_cohort}{An integer between 2 and 4.
If 2, then only comb and SoC.
If 3, then only Comb, Mono and SoC.
If 4, then all four arms.}
}
\value{
List containing: Responses and patients on experimental and control arm, total treatment successes and failures and final p-value
}
\description{
Simulates the cohort trial.
}
\examples{

random <- TRUE

rr_comb <- c(0.20, 0.25, 0.3)
prob_comb_rr <- c(0.4, 0.4, 0.2)
rr_plac <- c(0.10, 0.12, 0.14)
prob_plac_rr <- c(0.25, 0.5, 0.25)

rr_transform <- list(function(x) {return(c(0.85*(1 - x), (1-0.85)*(1-x), (1-0.85)*x, 0.85*x))})
prob_rr_transform <- 1

cohorts_start <- 2
design_type <- "combination"
arms_per_cohort <- 2
cohorts_max <- 5
safety_prob <- 0
sharing_type <- "concurrent"
sr_drugs_pos <- 5
sr_first_pos <- FALSE
n_int <- 50
n_fin <- 100
stage_data <- TRUE
cohort_random <- 0.01
cohort_offset <- 0
cohorts_sim <- Inf
random_type <- "risk_difference"
missing_prob <- 0.2
cohort_fixed <- 5
bio_lag <- 12
hist_lag <- 48
accrual_type <- "fixed"
accrual_param <- 9
time_trend <- 0.001

# Vergleich Combo vs Plac
Bayes_Sup1 <- matrix(nrow = 1, ncol = 2)
Bayes_Sup1[1,] <- c(0.00, 0.95)
Bayes_Sup <- list(list(Bayes_Sup1),
                  list(Bayes_Sup1))

a <- simulate_trial(
n_int = n_int, n_fin = n_fin, random_type = random_type,
rr_comb = rr_comb, rr_mono = rr_mono, rr_back = rr_back, rr_plac = rr_plac,
rr_transform = rr_transform, random = random, prob_comb_rr = prob_comb_rr,
prob_mono_rr = prob_mono_rr, prob_back_rr = prob_back_rr, prob_plac_rr = prob_plac_rr,
stage_data = stage_data, cohort_random = cohort_random, cohorts_max = cohorts_max,
sr_drugs_pos = sr_drugs_pos, sharing_type = sharing_type,
safety_prob = safety_prob, Bayes_Sup = Bayes_Sup, prob_rr_transform = prob_rr_transform,
cohort_offset = cohort_offset, sr_first_pos = sr_first_pos,
missing_prob = missing_prob, cohort_fixed = cohort_fixed, accrual_type = accrual_type,
accrual_param = accrual_param, bio_lag = bio_lag, hist_lag = hist_lag,
time_trend = time_trend, cohorts_start = cohorts_start, cohorts_sim = cohorts_sim
)

}
