% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cats.R
\name{summary.cats}
\alias{summary.cats}
\title{Summary of a cats object}
\usage{
\method{summary}{cats}(object, ...)
}
\arguments{
\item{object}{a cats object}

\item{...}{Dynamic parameter for the values of additional parameters
for the print method}
}
\description{
Summary of a cats object
}
\examples{
x <- cats()
summary(x)
}
\seealso{
cats
}
