% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercats.R
\name{super.cats}
\alias{super.cats}
\title{Generate supercats}
\usage{
super.cats(RR, MAFmax = 0.5, MAFmin = 0.005, by = 50, rep = 1536,
  SNPs = 1e+06, ncases, ncontrols, ncases2, ncontrols2,
  alpha = 0.05/SNPs, ...)
}
\arguments{
\item{RR}{description}

\item{MAFmax}{description}

\item{MAFmin}{description}

\item{by}{description}

\item{rep}{description}

\item{SNPs}{description}

\item{ncases}{description}

\item{ncontrols}{description}

\item{ncases2}{description}

\item{ncontrols2}{description}

\item{alpha}{description}

\item{...}{description}
}
\description{
Generate supercats
}
