% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tester.R
\name{tester}
\alias{tester}
\title{Tester}
\usage{
tester(age = 66, size = 4)
}
\arguments{
\item{age}{Age at diagnosis}

\item{size}{Size of tumour}
}
\description{
This function allows you to express your love of cats.
}
\examples{
tester()
}
\keyword{cats}
