% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_ocs.R
\name{trial_ocs}
\alias{trial_ocs}
\title{Calculates the operating characteristics of the cohort trial}
\usage{
trial_ocs(
  iter,
  coresnum = 1,
  save = FALSE,
  path = NULL,
  filename = NULL,
  ret_list = FALSE,
  ret_trials = FALSE,
  plot_ocs = FALSE,
  export = NULL,
  ...
)
}
\arguments{
\item{iter}{Number of program simulations that should be performed}

\item{coresnum}{How many cores should be used for parallel computing}

\item{save}{Indicator whether simulation results should be saved in an Excel file}

\item{path}{Path to which simulation results will be saved; if NULL, then save to current path}

\item{filename}{Filename of saved Excel file with results; if NULL, then name will contain design parameters}

\item{ret_list}{Indicator whether function should return list of results}

\item{ret_trials}{Indicator whether individual trial results should be saved as well}

\item{plot_ocs}{Should OCs stability plots be drawn?}

\item{export}{Should any other variables be exported to the parallel tasks?}

\item{...}{All other design parameters for chosen program}
}
\value{
List containing: Responses and patients on experimental and control arm, total treatment successes and failures and final p-value
}
\description{
Given the trial specific design parameters, performs a number
of simulations of the trial and saves the result in an Excel file
}
\examples{

random <- TRUE

rr_comb <- c(0.20, 0.25, 0.3)
prob_comb_rr <- c(0.4, 0.4, 0.2)
rr_plac <- c(0.10, 0.12, 0.14)
prob_plac_rr <- c(0.25, 0.5, 0.25)

rr_transform <- list(function(x) {return(c(0.85*(1 - x), (1-0.85)*(1-x), (1-0.85)*x, 0.85*x))})
prob_rr_transform <- 1

cohorts_start <- 2
design_type <- "combination"
arms_per_cohort <- 2
cohorts_max <- 3
safety_prob <- 0
sharing_type <- "concurrent"
sr_drugs_pos <- 5
sr_first_pos <- FALSE
n_int <- 50
n_fin <- 100
stage_data <- TRUE
cohort_random <- 0.1
cohort_offset <- 0
cohorts_sim <- Inf
random_type <- "risk_difference"
missing_prob <- 0.2
cohort_fixed <- NULL
bio_lag <- 12
hist_lag <- 48
accrual_type <- "fixed"
accrual_param <- 9
time_trend <- 0.001

# Vergleich Combo vs Plac
Bayes_Sup1 <- matrix(nrow = 1, ncol = 2)
Bayes_Sup1[1,] <- c(0.00, 0.95)
Bayes_Sup <- list(list(Bayes_Sup1),
                  list(Bayes_Sup1))

ocs <- trial_ocs(
n_int = n_int, n_fin = n_fin, random_type = random_type,
rr_comb = rr_comb, rr_mono = rr_mono, rr_back = rr_back, rr_plac = rr_plac,
rr_transform = rr_transform, random = random, prob_comb_rr = prob_comb_rr,
prob_mono_rr = prob_mono_rr, prob_back_rr = prob_back_rr, prob_plac_rr = prob_plac_rr,
stage_data = stage_data, cohort_random = cohort_random, cohorts_max = cohorts_max,
sr_drugs_pos = sr_drugs_pos, sharing_type = sharing_type,
safety_prob = safety_prob, Bayes_Sup = Bayes_Sup, prob_rr_transform = prob_rr_transform,
cohort_offset = cohort_offset, sr_first_pos = sr_first_pos,
missing_prob = missing_prob, cohort_fixed = cohort_fixed, accrual_type = accrual_type,
accrual_param = accrual_param, bio_lag = bio_lag, hist_lag = hist_lag,
time_trend = time_trend, cohorts_start = cohorts_start, cohorts_sim = cohorts_sim,
iter = 10, coresnum = 1, save = FALSE, ret_list = TRUE, plot_ocs = TRUE
)

ocs[[3]]

}
