% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_function.R
\name{cat_function}
\alias{cat_function}
\title{Expresses a well known distaste for Cats}
\usage{
cat_function(love = F)
}
\arguments{
\item{love}{defaults to FALSE for obvious reasons}
}
\description{
Expresses a well known distaste for Cats
}
\examples{
cat_function()
}
