% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_function.R
\name{cat_function}
\alias{cat_function}
\title{Cat loving function}
\usage{
cat_function(love = TRUE)
}
\arguments{
\item{love}{Do you love cats? Default is TRUE}
}
\value{

}
\description{
This function allows tou to love blue cats and add som unicorns in your coffee.
}
\examples{
cat_function(love=TRUE)
}
