% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATS.R
\name{getSignatureGenes}
\alias{getSignatureGenes}
\title{STEP 0: Separates the signature gene expression and the rest of the gene
expression}
\usage{
getSignatureGenes(signature, geneExp)
}
\arguments{
\item{signature}{Data frame formatted signature gene list}

\item{geneExp}{Gene by sample expression matrix. The first column of this
matrix should be gene names.}
}
\value{
Returns signature and rest of the genes' expression matrix as a
       list
}
\description{
STEP 0: Separates the signature gene expression and the rest of the gene
expression
}
\examples{
step0 <- getSignatureGenes(signature, gene_expression)

}
