% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_PPIN}
\alias{human_PPIN}
\title{Human PPI network}
\format{
Not Available
}
\source{
\url{https://doi.org/10.1038/s41540-017-0005-4}
}
\usage{
human_PPIN
}
\description{
Human PPI network
}
\examples{
data(human_PPIN, package = 'CATS')
}
\keyword{datasets}
