% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_function.R
\name{random_function}
\alias{random_function}
\title{A Random Function}
\usage{
random_function()
}
\description{
This function creates a tibble of standard normal values
}
\examples{
random_function()
}
\keyword{normal}
\keyword{random}
